requirements.txt
transformers
selenium
webdriver-manager
Pillow
pytesseract
scikit-learn
python-dotenv
